use double instead of float

+++ src/activation_functions.h
@@ -5,14 +5,14 @@
 #include <functional>
 
 struct ActivationFunctions {
-  static const std::function<float(float)> LOGISTIC;
-  static const std::function<float(float)> MUL_BY_TWO;
+  static const std::function<double(double)> LOGISTIC;
+  static const std::function<double(double)> MUL_BY_TWO;
 };
 
-const std::function<float(float)> ActivationFunctions::LOGISTIC =
-    [](float x) -> float { return 1.0 / (1.0 + exp(-1.0 * x)); };
+const std::function<double(double)> ActivationFunctions::LOGISTIC =
+    [](double x) -> double { return 1.0 / (1.0 + exp(-1.0 * x)); };
 
-const std::function<float(float)> ActivationFunctions::MUL_BY_TWO =
-    [](float x) -> float { return x * 2; };
+const std::function<double(double)> ActivationFunctions::MUL_BY_TWO =
+    [](double x) -> double { return x * 2; };
 
 #endif  // ACTIVATION_FUNCTIONS_H
